fix(admin-ui): an unauthenticated 302 carries an empty body — a pre-render session guard answers before Leptos renders
The router assembly moves from the binary into the lib
(ferroehr_admin_ui::server::router) so the integration tests drive the REAL
service — session layer, the new login_guard, the Leptos render, the
security headers — with oneshot, the same testable-lib doctrine the API
binary follows. login_guard refuses an unauthenticated GET/HEAD document
request to any guarded path with 302 → /login and an EMPTY body, before any
render work: previously the redirect carried the whole rendered console,
including the serialized failure diagnostics of every screen server fn, and
every anonymous hit paid a full server render. The public surface (/login,
/auth/*, /pkg/*, /api/*, favicon) is pinned by a unit test; the in-view
guard (AppShell's <Redirect/>) stays as the client-side navigation gate,
pinned by its own bare-render test — the Outlet-outside-Suspense hydration
rule is untouched. The guard test now asserts the empty body through the
routed service.
